Refactor worker and Enqueuer package
Keep Enqueuer public API
