+++ test/redis_queue_test.exs
@@ -3,7 +3,7 @@ Code.require_file "test_helper.exs", __DIR__
 defmodule JobQueueTest do
   use ExUnit.Case
   use Timex
-  alias Exq.Enqueuer.Server, as: Enqueuer
+  alias Exq.Enqueuer
   alias Exq.Redis.JobQueue
 
   setup_all do
@@ -70,8 +70,8 @@ defmodule JobQueueTest do
     JobQueue.enqueue_at(:testredis, "test", "default", time4, "MyWorker", [])
     time5 = Time.add(now, Time.from(300, :secs))
 
-    assert Enqueuer.queue_size(:testredis, "test", "default") == "0"
-    assert Enqueuer.queue_size(:testredis, "test", :scheduled) == "5"
+    assert Exq.Enqueuer.Server.queue_size(:testredis, "test", "default") == "0"
+    assert Exq.Enqueuer.Server.queue_size(:testredis, "test", :scheduled) == "5"
 
     assert JobQueue.scheduler_dequeue(:testredis, "test", ["default"], JobQueue.time_to_score(time2a)) == 2
     assert JobQueue.scheduler_dequeue(:testredis, "test", ["default"], JobQueue.time_to_score(time2b)) == 0
@@ -80,8 +80,8 @@ defmodule JobQueueTest do
     assert JobQueue.scheduler_dequeue(:testredis, "test", ["default"], JobQueue.time_to_score(time4)) == 1
     assert JobQueue.scheduler_dequeue(:testredis, "test", ["default"], JobQueue.time_to_score(time5)) == 1
 
-    assert Enqueuer.queue_size(:testredis, "test", "default") == "5"
-    assert Enqueuer.queue_size(:testredis, "test", :scheduled) == "0"
+    assert Exq.Enqueuer.Server.queue_size(:testredis, "test", "default") == "5"
+    assert Exq.Enqueuer.Server.queue_size(:testredis, "test", :scheduled) == "0"
 
     assert elem(JobQueue.dequeue(:testredis, "test", "default"), 0) != :none
     assert elem(JobQueue.dequeue(:testredis, "test", "default"), 0) != :none
